requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
Brotli==1.1.0
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cssselect2==0.8.0
curlify==2.2.1
Django==5.1.7
django-autocomplete-light==3.12.1
django-cors-headers==4.7.0
django-filter==25.1
django-rest==0.8.7
djangorestframework==3.15.2
dnspython==2.7.0
drf-writable-nested==0.7.2
et_xmlfile==2.0.0
fonttools==4.56.0
idna==3.10
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
prettytable==3.16.0
pycparser==2.22
pydyf==0.11.0
pymongo==4.11.3
PyMuPDF==1.25.4
PyPDF2==3.0.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
weasyprint==65.0
webencodings==0.5.1
zopfli==0.2.3.post1
